ContactList: Prevent Dialer from loading when Contacts app is launched.
This change prevents Dialer from being setup as the current tab when
Contacts is launched. This improves performance by preventing the
TwelveKeyDialer class from getting loaded when Contacts List is viewed.
When the Phone app is launched, the original behavior is unchanged and
the Dialer is set up as the current tab.

Change-Id: I67bb2f800f3b7cf339a160039a540734e1a953ed
CRs-Fixed: 228919