packages/apps/Contact/ Added null check in EditContactActivity.
When a raw contact is added or modified/edited, the system looks for matching (overlapping) raw
contacts with which to aggregate it. If it finds a single match,it creates a new contact that
contains the two raw contacts. In this process it checking AccountType which is by default null
and trying to call compare function with it which is null pointer.

skip AccountType Check if AccountType is null

Change-Id: Id9a519664caf961d718a2459ecf9a036c162b017
Crs-Fixed: 247484