Contacts: Enable adding call from call log
CallDetailActivity and ContactsListActivity won't let a new call go through if
there's a currently active call. If the user were to dial an emergency number (or any
other number for that matter) InCallScreen will be brought up, but no new call
will be added, which can be inadequate particularly in the Emergency Call scenario.

To fix this:
- Use the same behavior used by the "Add New Call" button in InCallScreen, which will
bring up the Dialer with the number to dial, ready for the user to press the Send key
to actually dial the call
- React to Key up instead of key down (for consistency, and to prevent the matching
  key up event from being handled by the Dialer, which would immediately dial the call)

Change-Id: I3b32e21e485cc1ad393944106fb80f908bf3d0e2
CRs-Fixed: 257678