packages/apps/Contacts: Added dialpad to TwelveKeyDialer in landscape mode
Added dialpad to the TwelveKeyDialer In landscape mode
Before this change in landscape mode the only keypad
available is the Text Editor's default Dialerkeypad.
This is used to enter numbers while making a call.
With this change the TwelveKeyDialer's dialpad will
be used for dialing numbers similar to portrait mode.
Using this dialpad will allow the number key '1' to
be mapped to voicemail number

CRs-Fixed: 259120

Change-Id: Ica9c8668fa9bcade9b807e405af38fab0f18cfd8