packages/apps/Contacts: Added dialpad to TwelveKeyDialer in landscape mode
Added dialpad to the TwelveKeyDialer In landscape mode
Before this change in landscape mode the only keypad
available is the Text Editor's default Dialerkeypad.
This is used to enter numbers while making a call.
With this change the TwelveKeyDialer's dialpad will
be used for dialing numbers similar to portrait mode.
Using this dialpad will allow the number key '1' to
be mapped to voicemail number

Change-Id: I9f2a000b9983547f6501a2ba9008991b75fcc74d
CRs-Fixed: 269508