Contacts: Enable adding call from call log
CallDetailActivity and ContactsListActivity won't let a new
call go through if there's a currently active call. If the
user were to dial an emergency number (or any other number
for that matter) InCallScreen will be brought up, but no new
call will be added, which can be inadequate particularly in
the Emergency Call scenario.

To fix this:
- Use the same behavior used by the "Add New Call" button in
InCallScreen, which will bring up the Dialer with the
number to dial, ready for the user to press the Send key to
actually dial the call
- React to Key up instead of key down (for consistency, and
to prevent the matching  key up event from being handled by
the Dialer, which would immediately dial the call)

Change-Id: Ia241fe4e20863dd094fb175c7364fa662dfae65d
CRs-Fixed: 303527