packages/apps/Contacts: Dialpad changes for landscape mode
* Add dialpad to Dialer in landscape mode

Before this change in landscape mode the only keypad
available is the Text Editor's default Dialerkeypad.
This is used to enter numbers while making a call.
With this change the Phone Dialer's dialpad is
use for dialing numbers.
This is similar to portrait mode.
This dialpad allows the number key '1' to
be mapped to voicemail number

* Interpret keypress as numbers on ffa Phone app

Keypress from the hard keyboard on ffa is interpreted as
letters and not numbers in Phone app since soft dialpad
is absent.
Change input mode to phone always so that any keypad
digit entry will be interepreted as numbers in PhoneApp.

Change-Id: I64be9d3577837e1c0b04b1476723683913e8b27d
CRs-Fixed: 269508