Contacts: Prevent text box from showing old phone number
When dialing a number using the InputMethod's "Done" button, the previously dialed number
was briefly displayed even though the current number was properly dialed.
This was caused because dialButtonPressed was called twice per key: once for key down, which
cleared the numbers from the EditText, and once for key up, which made dialButtonPressed
believe this was a case in which it should show the old number (since no more digits were
in the EditText.

Trigger a call to dialButtonPressed only on key up.

CRs-Fixed: 328878
(cherry picked from commit 999108e67d7b715d492c0c1ec5e89aa1cd66eb96)

Change-Id: Ic428aed01dcab77b0dadcdc0f61e6eb76f7070fc