Contacts: Prevent text box from showing old phone number
When dialing a number using the InputMethod's "Done" button, the previously dialed number
was briefly displayed even though the current number was properly dialed.
This was caused because dialButtonPressed was called twice per key: once for key down, which
cleared the numbers from the EditText, and once for key up, which made dialButtonPressed
believe this was a case in which it should show the old number (since no more digits were
in the EditText.

Trigger a call to dialButtonPressed only on key up.

Change-Id: Id8d698bde3c37ba9205c179a4a65edeaffae24e7
CRs-Fixed: 328878