Contacts: Fix choose an account dialog display error in Contacts.
When we add accounts in Contacts, it will start "ChooseAccountTypeActivity",
and if user select SIM card/Phone item, it will just add account of SIM card/Phone. But the account of SIM card/Phone also exist.
So it will stay in the "ChooseAccountTypeActivity".

If the user select SIM card/Phone item, only cancel the add account request
because it isn't real account like google or corporate.

CRs-Fixed: 502085

Change-Id: I579329cecaf63c8e91df99695e9738577f0956fa