Contact: Can't set contact photo from FileExplorer
In contact app the RESULT_OK is defined as -1, but it is defined as 1
in FileExplorer. And there isn't function that set contact photo from
FileExplorer in contact app, this cause user can't set contact photo
from FileExplorer.

Define RESULT_OK as -1 in FileExplorer, and add the function that
set contact photo from FileExplorer in contact app.

Change-Id: I89d0ccc8e9f75371d5c8c2aaec577d9b94c04638
CRs-Fixed: 506082