Contacts: change AES encrypt and decrypt algorithm for baidu yun
When user set Auto sync to Baidu Yun, Contacts and BaiduNetdisk use the
AES encryption algorithm to encrypt and decrypt the status of Auto sync.
But the provider for secureRandom has been changed to OpenSSL instead of
Crypto on android 4.2. So the encrypt and decrypt method can't work
normally. And the status which has been encrypted is saved in
BaiduNetdisk's database. And then BaiduNetdisk will decrypt and encrpt
the status again.

Change the default provider for secureRandom to AES encryption algorithm
to make it can work on Android 4.2. But also need Baidu to change codes
to adapt current QRD platform.

CRs-Fixed: 510170

Change-Id: I3494f47f6c1a8893e7437c7ffc7352724b5eaf3d