Contact: Can't export to storage and share visible contacts
There isn't permission of write media to storage, so user can't
export contacts to storage. And it doesn't update the account to
visible when an account exist, so it return a incorrect value "0"
when get the value that judgment the contact is visible or not,
this cause the contact data can't be inserted into the visible
contact table, and user can't share visible contacts.

Add the permission of write media to storage and update the account
to visible when an account exist.

CRs-Fixed: 521691

Change-Id: Ice0c92641587eb0d73c31d0f34ff885ce78e0ea5