Contacts: Fix can't save contact after change SIM contact name
If we not change the phone field in the UI, this values' after
would not be changed, so it will be null. Now in the Sim build
diff, we should get the new phone number form the after, because
after is null, so we can not get the new phone number.

If we can not get new phone number, we would merge it from old
one. And do the same for name, anr and email.

CRs-Fixed: 521698

Change-Id: I52021cebdeee69c618aa676e82b4140a8d4f6e4b