Contacts: Can not add members to group
If contact is just removed during this session,it will be
added to list of members to remove, if contact was added, it
will be added to list of members to add. At the end, it will
add all the contacts in list of members to add and remove all
the contacts in list of members to remove.

If the contact was just removed during this session, remove it
from the list of members to remove, and do not need to add it to
the list of members to add.

CRs-Fixed: 535182

Change-Id: I429abae629b7f6cf72ae1900d40ea5a21d47a401