Contact: People crash when quickly tap "OK" several times
String objects are immutable, and need to create a new object in memory
each time we use it, so the efficiency is very low, and phone will crash
when we quickly tap "OK" several times.

StringBuilder is a dynamic object, and it allow expand the length of
the string. We needn't to create a new object in memory each time
we use it, and it can improve the efficiency greatly. So we use it
instead of String.

CRs-Fixed: 537159

Change-Id: I2e60deb70c449ba0510d48dd6b6986b46b4c524f