Contacts: Phone crash when add contacts to local group
The program run slowly when add too many contacts to local group
once, in this time if do other operations about database, Exception
will happen. And the "OK" button should not be clicked to add
contacts to group when the count of selected contacts is zero.

Reduce the limit of contacts to 500, and make the "OK" button unable
be clicked when the count of selected contacts is zero.

CRs-Fixed: 556021

Change-Id: Ib924c6cbc3630029a8cdec166439fe89f6b3d8fc