Contacts: Cancel the system outgoing animation while exiting
The system will perform an outgoing animation while finishing activity,
and it will display a black background during the animation, and the
background of quick contact activity is transparent, so it look likes
there is a black screen.

Don't use the system outgoing animation for the quick contact activity.

CRs-Fixed: 674524

Change-Id: If4ea44ddacd868b03dbbd886d0263ec376c1ac38