Contacts: "No contact" always show after delete all contacts
After delete all contacts, the MultiPickContactActivity has been
destroyed. Sometimes the mQueryHandler still continues querying
what makes the adapter is still been used and not been destroyed,
so the "No contact" always show.

When the activity has been destroyed, set the adapter be null so that
the adapter can be destroyed with the activity.

Change-Id: I10835d1673b51a1b4c1ccd0a771a1db6c79b6554
CRs-Fixed: 844501