Contacts: Can't add vcard as MMS attachment
When add one contact as MMS attachment, the contact's
lookup uri contains ":". When MMS quering the contact
in contact's DB, because of the ":" as multi lookup
uri's separator in contact's DB, so the contact's lookup
uri will be split, this will cause MMS can not find the
contact in contact's DB.

Because of some contact which belong to hotmail account's
lookup uri contains ":", the lookup uri will be split even
only one lookup uri . so use the "&" to replace the ":" as
the multi lookup uri's separator, then only multi lookup uri
will be split.

Change-Id: I6e6e1cf65aa9bf31b650a8a8aa57d275fb164e7f
CRs-Fixed: 848386